+++ plugins/modules/zhmc_adapter.py
@@ -345,12 +345,13 @@ adapter:
 
 import logging  # noqa: E402
 import traceback  # noqa: E402
-from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule, \
+    missing_required_lib  # noqa: E402
 
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, ParameterError, to_unicode, \
-    process_normal_property, eq_hex, missing_required_lib, \
-    common_fail_on_import_errors, parse_hmc_host, blanked_params  # noqa: E402
+    process_normal_property, eq_hex, common_fail_on_import_errors, \
+    parse_hmc_host, blanked_params  # noqa: E402
 
 try:
     import urllib3