+++ plugins/modules/zhmc_hba.py
@@ -241,11 +241,12 @@ hba:
 
 import logging  # noqa: E402
 import traceback  # noqa: E402
-from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule, \
+    missing_required_lib  # noqa: E402
 
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, ParameterError, wait_for_transition_completion, \
-    eq_hex, to_unicode, process_normal_property, missing_required_lib, \
+    eq_hex, to_unicode, process_normal_property, \
     common_fail_on_import_errors, parse_hmc_host, blanked_params  # noqa: E402
 
 try: