+++ plugins/modules/zhmc_lpar.py
@@ -615,14 +615,14 @@ lpar:
 
 import logging  # noqa: E402
 import traceback  # noqa: E402
-from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule, \
+    missing_required_lib  # noqa: E402
 
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, ParameterError, StatusError, \
     ensure_lpar_inactive, ensure_lpar_active, ensure_lpar_loaded, to_unicode, \
-    process_normal_property, missing_required_lib, \
-    common_fail_on_import_errors, pull_properties, parse_hmc_host, \
-    blanked_params, removed_dict  # noqa: E402
+    process_normal_property, common_fail_on_import_errors, pull_properties, \
+    parse_hmc_host, blanked_params, removed_dict  # noqa: E402
 
 try:
     import urllib3