+++ plugins/modules/zhmc_partition.py
@@ -619,15 +619,15 @@ import uuid
 import random
 import types
 from operator import itemgetter  # noqa: E402
-from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule, \
+    missing_required_lib  # noqa: E402
 
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, ParameterError, StatusError, stop_partition, \
     start_partition, wait_for_transition_completion, eq_hex, to_unicode, \
-    process_normal_property, missing_required_lib, ImageError, \
-    common_fail_on_import_errors, pull_properties, parse_hmc_host, \
-    blanked_params, removed_dict, UNKNOWN_NAME, object_from_uri, \
-    object_properties  # noqa: E402
+    process_normal_property, ImageError, common_fail_on_import_errors, \
+    pull_properties, parse_hmc_host, blanked_params, removed_dict, \
+    UNKNOWN_NAME, object_from_uri, object_properties  # noqa: E402
 
 try:
     import urllib3