+++ plugins/modules/zhmc_user_pattern.py
@@ -286,11 +286,12 @@ user_pattern:
 import uuid  # noqa: E402
 import logging  # noqa: E402
 import traceback  # noqa: E402
-from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule, \
+    missing_required_lib  # noqa: E402
 
 from ..module_utils.common import log_init, open_session, close_session, \
     hmc_auth_parameter, Error, ParameterError, to_unicode, \
-    process_normal_property, missing_required_lib, underscore_properties, \
+    process_normal_property, underscore_properties, \
     common_fail_on_import_errors, parse_hmc_host, blanked_params  # noqa: E402
 
 try: